修改数据类型从long到int，中央空调初始化

+++ backend/src/main/java/com/backend/pojo/ACServiceObject.java
@@ -30,7 +30,7 @@ public class ACServiceObject {
     /**
      * 设定风速
      */
-    private long speedLevel;
+    private int speedLevel;
     /**
      * 目标温度
      */
